fixed location error, breadcrumbs and component refactoring
