+++ litmus-portal/frontend/src/components/Sections/CreateWorkflow/WeightSlider/index.tsx
