+++ litmus-portal/frontend/src/components/Sections/CreateWorkflow/WeightSlider/styles.ts
